more compiler warning cleanup
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@33349 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ viterbi/connect.cc
@@ -361,7 +361,7 @@ const OutList& Connect::flatten(OutList& alternatives)
 			return set->get_outgoing_set();
 	}
 
-	for (int i = 0; i < asize; i++)
+	for (size_t i = 0; i < asize; i++)
 	{
 		Set* set = dynamic_cast<Set*>(alternatives[i]);
 		if (set and (1 == set->get_arity()))
